// types/pf2e/src/module/actor/character/data.d.ts
interface CharacterStrike extends StrikeData {
    item: WeaponPF2e<CharacterPF2e>;
    /** Whether this attack is visible on the sheet */
    visible: boolean;
    /** Domains/selectors from which modifiers are drawn */
    domains: string[];
    /** Whether the character has sufficient hands available to wield this weapon or use this unarmed attack */
    handsAvailable: boolean;
    altUsages: CharacterStrike[];
    auxiliaryActions: WeaponAuxiliaryAction[];
    weaponTraits: TraitViewData[];
    versatileOptions: VersatileWeaponOption[];
}
interface VersatileWeaponOption {
    value: DamageType;
    selected: boolean;
    label: string;
    glyph: string | null;
}